% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.IFS_S3}
\alias{plot.IFS_S3}
\title{Plot fractal calculated by IFS method}
\usage{
\method{plot}{IFS_S3}(IFS, point = c(0, 0), n)
}
\arguments{
\item{IFS}{object of class IFS_FS}

\item{point}{starting point}

\item{n}{number of iteration}
}
\value{
Creates plot of fractal
}
\description{
Plot fractal calculated by IFS method
}
\examples{
# Example sierpinski_triangle from createIFS
plot(sierpinski_triangle, n = 10)
}
